gadget-core: Make `#![no_std]` compatible
