+++ blueprint-test-utils/src/test_ext.rs
@@ -33,7 +33,8 @@ use gadget_common::tangle_runtime::api::services::calls::types::register::{
 use gadget_common::tangle_subxt::subxt::ext::subxt_core::utils;
 use gadget_common::tangle_subxt::subxt::{OnlineClient, SubstrateConfig};
 use gadget_common::Error;
-use gadget_core::job_manager::{ProtocolMessageMetadata, SendFuture, WorkManagerInterface};
+use gadget_core::job::SendFuture;
+use gadget_core::job::protocol::{ProtocolMessageMetadata, WorkManagerInterface};
 use libp2p::Multiaddr;
 use sp_application_crypto::ecdsa;
 use sp_core::{sr25519, Pair};
