+++ core/src/gadget/error.rs
@@ -2,12 +2,15 @@ use core::error::Error;
 use core::fmt::{Debug, Display, Formatter};
 
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub struct SubstrateGadgetError {}
+pub enum GadgetError {
+    FinalityNotificationStreamEnded,
+    ProtocolMessageStreamEnded,
+}
 
-impl Display for SubstrateGadgetError {
+impl Display for GadgetError {
     fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
         Debug::fmt(self, f)
     }
 }
 
-impl Error for SubstrateGadgetError {}
+impl Error for GadgetError {}
